chore(MFLP-58): Reflect Feedback
- totalPrice long -> BigDecimal로 변경
- merchantUid 필드에 unique 조건 추가
- Checked Exception 인식을 위해  @Transactional(rollbackFor = Exception.class) 추가

+++ src/main/java/api/buyhood/domain/order/dto/response/ApplyOrderRes.java
@@ -3,6 +3,8 @@ package api.buyhood.domain.order.dto.response;
 import api.buyhood.domain.cart.dto.response.CartRes;
 import api.buyhood.domain.order.enums.OrderStatus;
 import api.buyhood.domain.order.enums.PaymentMethod;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,13 +17,13 @@ public class ApplyOrderRes {
 
 	private final Long storeId;
 	private final CartRes orderInfo;
-	private final long totalPrice;
+	private final BigDecimal totalPrice;
 	private final OrderStatus status;
 	private final String requestMessage;
 	private final LocalDateTime createAt;
 
 
-	public static ApplyOrderRes of(Long storeId, CartRes orderInfo, long totalPrice,
+	public static ApplyOrderRes of(Long storeId, CartRes orderInfo, BigDecimal totalPrice,
 		OrderStatus status, LocalDateTime createAt, String requestMessage) {
 		return ApplyOrderRes.builder()
 			.storeId(storeId)
